LayoutTests/media/track/in-band/track-in-band-srt-mkv-addtrack-expected.txt

Testing text track "addtrack" events.

EVENT(addtrack)
EXPECTED (event.track == 'video.textTracks[0]') OK

EVENT(addtrack)
EXPECTED (event.track == 'video.textTracks[1]') OK

EVENT(canplaythrough)
END OF TEST
